feat(lib/dataplane): extract reusable primitives from dataplane
Lifts eight helpers from dataplane/{dataplane,worker}.c into
lib/dataplane/ so an in-process harness (next commit) can compose
them without copy-paste.

Adds dataplane_time_ns_fn override hook in lib/dataplane/time/clock.h
so tests can install a deterministic clock.

Production behaviour deltas (both align with mock): dlsym module and
device loaders free the returned struct on success; topology
allocation checks OOM.
