+++ lib/dataplane/config/meson.build
@@ -1,9 +1,17 @@
 dependencies = [
   lib_module_dp_dep,
   lib_pipeline_dp_dep,
+  lib_errors_dep,
+  lib_logging_dep,
+  lib_counters_dep,
 ]
 
 sources = files(
+  'agent.c',
+  'bootstrap.c',
+  'counter_storage.c',
+  'module_loader.c',
+  'topology.c',
   'zone.c',
 )
