+++ lib/dataplane/worker/meson.build
@@ -1,12 +1,19 @@
 dependencies = [
   lib_common_dep,
+  lib_counters_dep,
+  lib_logging_dep,
+  lib_errors_dep,
   lib_config_dp_dep,
+  lib_pipeline_dp_dep,
+  lib_module_dp_dep,
   libdpdk_inc_dep,
 ]
 
 
 sources = files(
+  'pipeline_round.c',
   'worker.c',
+  'counters.c',
 )
 
 lib_worker_dp = static_library(
